% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpCleaner.R
\name{clean_sdp}
\alias{clean_sdp}
\title{Clean school/dept/program values}
\usage{
clean_sdp(courses, course_code_colname = "code")
}
\arguments{
\item{courses}{(tibble)}

\item{course_code_colname}{(character)}
}
\description{
Clean school/dept/program values
}
