% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpCleaner.R
\name{id_schools}
\alias{id_schools}
\title{Join school IDs to courses by school names}
\usage{
id_schools(courses, school_ids)
}
\arguments{
\item{courses}{(tibble)}

\item{school_ids}{(tibble)}
}
\description{
Join school IDs to courses by school names
}
