% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpCleaner.R
\name{rename_depts}
\alias{rename_depts}
\title{Rename courses' department names}
\usage{
rename_depts(courses, dept_renames)
}
\arguments{
\item{courses}{(tibble)}

\item{dept_renames}{(tibble)}
}
\description{
Rename courses' department names
}
