% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpCleaner.R
\name{rename_schools}
\alias{rename_schools}
\title{Rename courses' school names}
\usage{
rename_schools(courses, school_renames)
}
\arguments{
\item{courses}{(tibble)}

\item{school_renames}{(tibble)}
}
\description{
Rename courses' school names
}
